+++ Src/Engine/Pipeline.h
@@ -15,7 +15,7 @@ public:
     virtual ~IPipeline() = default;
 
     void create(vk::CDevice::CPtr vDevice, VkRenderPass vRenderPass, VkExtent2D vExtent, uint32_t vSubpass = 0);
-    void setImageNum(size_t vImageNum);
+    void setImageNum(uint32_t vImageNum);
     void destroy();
     void bind(VkCommandBuffer vCommandBuffer, size_t vImageIndex);
     bool isValid() const { return m_Pipeline != VK_NULL_HANDLE; }
